assembly block impl
